add drag helper module and tests
